refactor: reorg all routes
